+++ HISTORY.md
@@ -10,6 +10,7 @@
 
 ### 2021-04-23
 * store: serialize sprite-order from scenes
+* gui: sceneified refreshIDE()
 
 ### 2021-04-22
 * store, gui: first pass at deserializing multi-scene projects